hotfix de la vista de scout y scouter
